requirements.txt
beautifulsoup4==4.9.1
python-kasa==0.4.0.dev1
requests==2.23.0
RPi.GPIO==0.7.0
google-cloud-texttospeech==2.2.0